// src/app/service/model/beneficiario.model.ts
export interface BeneficiarioModel {
  idBeneficiario:    number;
  nome:              string;
  cpf:               string;
  telefone:          string;
  endereco:          string;
  numeroCarteirinha: string;
  ativo:             boolean;
  email:             string;
  senha:             string;
}